biblegpt_project/requirements.txt
flask
pandas
flask-cors
python-dotenv
google-generativeai